feat - typedef all the int and uint values (i8, i16, i32, u8, u16...) and modified in the code accordingly. - add welcome and warning prints. Issue #2
func
- add fn_getline() to read the data.psv file line by line. This fixed issue #1
	- got also rid of strtok(). Issue #2
- add replace_all() to string.c file
	- Now all the pipes will be fixed to '!'. Issue #2

other:
- no more id's saved on datafile. issue #2
- Reordered some functions

issues:
- #1 fixed. Read more on func section.
- #2 The final bullet with README.md is not done. All the others fixed. See above.

+++ src/data.c
@@ -1,6 +1,6 @@
 #include "todo.h"
 
-int init_data(s_data* data)
+i8 init_data(s_data* data)
 {
 	data->items = malloc(10 * sizeof(s_item));
 	if (data->items == NULL) {
@@ -12,7 +12,7 @@ int init_data(s_data* data)
 	return 0;
 }
 
-int resize_data(s_data* data)
+i8 resize_data(s_data* data)
 {
 	data->capacity *= 2;
 	data->items = realloc(data->items, data->capacity * sizeof(s_item));
